+++ subxt/src/lib.rs
@@ -155,7 +155,6 @@ pub mod constants;
 pub mod dynamic;
 pub mod error;
 pub mod events;
-pub mod extrinsics;
 pub mod metadata;
 pub mod rpc;
 pub mod runtime_api;
